third_party/blink/web_tests/wpt_internal/environment-integrity/navigator-environment-integrity.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
  'use strict';

  import { MockEnvironmentIntegrity } from "./resources/mock.js";

  const mockEnvironmentIntegrity = new MockEnvironmentIntegrity();

  test(() => {
    assert_idl_attribute(navigator, 'getEnvironmentIntegrity');
  }, 'The getEnvironmentIntegrity method should exist on the navigator object');

  promise_test(async () => {
    try {
      await navigator.getEnvironmentIntegrity('contentBinding');
    } catch (exception) {
      assert_unreached();
    }
  }, 'The getEnvironmentIntegrity method can be called');

  // TODO(crbug.com/1439945): Temporary test to confirm our mocks work until we can actually start
  // using them once we make use of the getEnvironmentIntegrity mojo call.
  promise_test(async () => {
    await navigator.getEnvironmentIntegrity('contentBinding');
    assert_true(mockEnvironmentIntegrity.wasCalled);
  }, 'Mock environment integrity should be called');
</script>
